requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic-settings
openai
google-generativeai
python-multipart
